Move DependencyController to controller package
